requirements.txt
# Core dependencies
Flask==2.3.3
geopandas==0.14.0
shapely==2.0.2
requests==2.31.0

# Geospatial dependencies
fiona==1.9.5
pyproj==3.6.1
rtree==1.1.0

# Web framework dependencies
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Data processing
pandas==2.1.3
numpy==1.25.2

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
flake8==6.1.0
black==23.10.1

# Production dependencies
gunicorn==21.2.0
python-dotenv==1.0.0